# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Araponga API",
    "version": "v1",
    "description": "Araponga é uma plataforma comunitária orientada a território. Esta API expõe recursos de território, feed comunitário, entidades do mapa e saúde do território. O objetivo do MVP é viabilizar cadastro, visualização e curadoria comunitária com segurança e governança mínima.",
    "contact": {
      "name": "Araponga (maintainers)"
    },
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/api/v1/platform-fees": {
      "get": {
        "tags": [
          "PlatformFees"
        ],
        "summary": "Lista configurações ativas de fee por território.",
        "parameters": [
          {
            "name": "territoryId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlatformFeeResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "PlatformFees"
        ],
        "summary": "Atualiza ou cria uma configuração de fee.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformFeeRequest"
              },
              "examples": {
                "configurarFee": {
                  "summary": "Configurar fee do território",
                  "value": {
                    "territoryId": "22222222-2222-2222-2222-222222222222",
                    "listingType": "PRODUCT",
                    "feeMode": "PERCENTAGE",
                    "feeValue": 0.05,
                    "currency": "BRL",
                    "isActive": true
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformFeeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PlatformFeeRequest": {
        "required": [
          "territoryId",
          "listingType",
          "feeMode",
          "feeValue",
          "isActive"
        ],
        "type": "object",
        "properties": {
          "territoryId": {
            "type": "string",
            "format": "uuid"
          },
          "listingType": {
            "type": "string"
          },
          "feeMode": {
            "type": "string"
          },
          "feeValue": {
            "type": "number",
            "format": "decimal"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "PlatformFeeResponse": {
        "required": [
          "id",
          "territoryId",
          "listingType",
          "feeMode",
          "feeValue",
          "isActive",
          "createdAtUtc",
          "updatedAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "territoryId": {
            "type": "string",
            "format": "uuid"
          },
          "listingType": {
            "type": "string"
          },
          "feeMode": {
            "type": "string"
          },
          "feeValue": {
            "type": "number",
            "format": "decimal"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}